refactor: reorderをutilsに分離
